+++ covid-data/src/main/java/com/moshenskyi/feature_covid_data/network/CovidRepository.kt
@@ -1,4 +1,4 @@
-package com.moshenskyi.feature_covid_data.domain_entity
+package com.moshenskyi.feature_covid_data.network
 
 import com.moshenskyi.feature_covid_data.model.general.CountryRecordsItem
 import com.moshenskyi.feature_covid_data.model.vaccine.VaccineCapacityInfoItem
